+++ src/model/KNode.ts
@@ -257,18 +257,6 @@ export class KNode {
     return false
   }
 
-  /**
-   * 【API】判断节点是否不可编辑的，如果是返回设置不可编辑的那个节点，否则返回null
-   */
-  getUneditable(): KNode | null {
-    if (this.hasMarks() && this.marks!['contenteditable'] == 'false') {
-      return this
-    }
-    if (!this.parent) {
-      return null
-    }
-    return this.parent.getUneditable()
-  }
 
   /**
    * 【API】当前节点是否只包含占位符